catchat-backend/requirements.txt
fastapi
uvicorn
openai
python-dotenv
pydantic
mysql-connector-python
